==> notebooks/4. generate_latex.ipynb <==
import pandas as pd
from tabulate import tabulate
from pathlib import Path
import json

# Set the default year to 2023
year_ = "2023"

output_directory = Path(f"../miccai{year_}")

### Inter-rater reliability

df_inter = pd.read_csv(output_directory / f"stats_inter_rater/inter_rater_stats.csv", sep= "\t")
print(tabulate(df_inter, headers='keys', tablefmt='psql'))

latex_directory = output_directory / "latex"
if not latex_directory.is_dir():
    latex_directory.mkdir()

with open(latex_directory / "table_inter_rater.tex", 'w') as f_:
    print("\\begin{tabular}{lll}", file=f_)
    print("\\hline", file=f_)
    print(f"& $\kappa$ & 95\% CI\\\\", file=f_)
    print("\\hline", file=f_)
    c_=df_inter.iloc[0].to_numpy()
    print(f"Models and algorithms & {c_[1]:.2f} & [{c_[2]:.2f}, {c_[3]:.2f}]  \\\\", file=f_)
    c_=df_inter.iloc[1].to_numpy()
    print(f"Datasets & {c_[1]:.2f} & [{c_[2]:.2f}, {c_[3]:.2f}]  \\\\", file=f_)
    c_=df_inter.iloc[2].to_numpy()
    print(f"Code & {c_[1]:.2f} & [{c_[2]:.2f}, {c_[3]:.2f}]  \\\\", file=f_)
    c_=df_inter.iloc[3].to_numpy()
    print(f"Experimental results & {c_[1]:.2f} & [{c_[2]:.2f}, {c_[3]:.2f}]  \\\\", file=f_)
    c_=df_inter.iloc[4].to_numpy()
    print(f"Error bars/stat. sign.& {c_[1]:.2f} & [{c_[2]:.2f}, {c_[3]:.2f}]  \\\\", file=f_)
    c_=df_inter.iloc[6].to_numpy()
    print(f"Statement & {c_[1]:.2f} & [{c_[2]:.2f}, {c_[3]:.2f}]  \\\\", file=f_)
    c_=df_inter.iloc[7].to_numpy()
    print(f"Comments & {c_[1]:.2f}  & [{c_[2]:.2f}, {c_[3]:.2f}]  \\\\", file=f_)
    c_=df_inter.iloc[8].to_numpy()
    print(f"Meta-category & {c_[1]:.2f}  & [{c_[2]:.2f}, {c_[3]:.2f}]  \\\\", file=f_)
    c_=df_inter.iloc[5].to_numpy()
    print(f"Code is or will be available & {c_[1]:.2f} & [{c_[2]:.2f}, {c_[3]:.2f}]  \\\\", file=f_)
    c_=df_inter.iloc[9].to_numpy()
    print(f"Code repository provided and not empty & {c_[1]:.2f} & [{c_[2]:.2f}, {c_[3]:.2f}]  \\\\", file=f_)
    print("\\hline", file=f_)
    print("\\end{tabular}", file=f_)

print(f"Wrote tex file {latex_directory / 'table_inter_rater.tex'}")

### Stats-rating category

with open(output_directory / "stats_rating" / "data.json", 'r') as file:
    data_json = json.load(file)

nb_reviews = data_json['nb_reviews']
print(f"Number of Papers: {data_json['nb_papers']}")
print(f"Number of Reviews: {nb_reviews}")

df_inter = pd.read_csv(output_directory / "stats_rating" / "1-category.csv", sep= ";")
print(tabulate(df_inter, headers='keys', tablefmt='psql'))

latex_directory = output_directory / "latex"
if not latex_directory.is_dir():
    latex_directory.mkdir()

with open(latex_directory / "table_checklist_categories.tex", 'w') as f_:
    print("\\begin{tabular}{llll}", file=f_)
    print("\\hline", file=f_)
    print(f"Category/item & \% & 95\% CI  & N \\\\", file=f_)
    print("\\hline", file=f_)
    c_=df_inter.iloc[0].to_numpy()
    print(f"Models and algorithms & {c_[2]:.0f}\% & [{c_[3]:.0f}\%, {c_[4]:.0f}\%] & {c_[1]:.0f}/{nb_reviews}\\\\", file=f_)
    c_=df_inter.iloc[1].to_numpy()
    print(f"Datasets & {c_[2]:.0f}\% & [{c_[3]:.0f}\%, {c_[4]:.0f}\%] & {c_[1]:.0f}/{nb_reviews}\\\\", file=f_)
    c_=df_inter.iloc[2].to_numpy()
    print(f"Code & {c_[2]:.0f}\% & [{c_[3]:.0f}\%, {c_[4]:.0f}\%] & {c_[1]:.0f}/{nb_reviews}\\\\", file=f_)
    c_=df_inter.iloc[3].to_numpy()
    print(f"Experimental results & {c_[2]:.0f}\% & [{c_[3]:.0f}\%, {c_[4]:.0f}\%] & {c_[1]:.0f}/{nb_reviews}\\\\", file=f_)
    c_=df_inter.iloc[4].to_numpy()
    print(f"Error bars/stat. sign.& {c_[2]:.0f}\% & [{c_[3]:.0f}\%, {c_[4]:.0f}\%] & {c_[1]:.0f}/{nb_reviews}\\\\", file=f_)
    print("\\hline", file=f_)
    print("\\end{tabular}", file=f_)

print(f"Wrote tex file {latex_directory / 'table_checklist_categories.tex'}")

### Agreement between reviewers (Fleiss' kappa)

df_inter = pd.read_csv(output_directory / "stats_rating" / "6-kappas.csv", sep= ";")
print(tabulate(df_inter, headers='keys', tablefmt='psql'))

latex_directory = output_directory / "latex"
if not latex_directory.is_dir():
    latex_directory.mkdir()

with open(latex_directory / 'table_agreement_reviewers.tex', 'w') as f_:
    print("\\begin{tabular}{lll}", file=f_)
    print("\\hline", file=f_)
    print(f" & $\kappa$ & 95\% CI  \\\\", file=f_)
    print("\\hline", file=f_)
    c_ = df_inter.iloc[0].to_numpy()
    print(f"Statement & {c_[1]:.2f} & [{c_[2]:.2f}, {c_[3]:.2f}] \\\\", file=f_)
    c_ = df_inter.iloc[2].to_numpy()
    print(f"Meta-categories & {c_[1]:.2f} & [{c_[2]:.2f}, {c_[3]:.2f}] \\\\", file=f_)
    print("\\hline", file=f_)
    print("\\end{tabular}", file=f_)

print(f"Wrote tex file {latex_directory / 'table_agreement_reviewers.tex'}")

### Code availability

df_here = pd.read_csv(
    output_directory / "stats_rating" / "7-code_number.csv",
    sep= ";",
    header = None,
    index_col = 0,
).transpose()
print(tabulate(df_here, headers='keys', tablefmt='psql', showindex= True))

df_here2 = pd.read_csv(
    output_directory / "stats_rating" / "8-code_percent.csv",
    sep= ";",
    header = None,
    index_col = 0,
).transpose()
print(tabulate(df_here2, headers='keys', tablefmt='psql'))

latex_directory = output_directory / "latex"
if not latex_directory.is_dir():
    latex_directory.mkdir()

with open(latex_directory / 'table_code_availability.tex', 'w') as f_:
    print("\\begin{tabular}{|l|l|l|l|}", file=f_)
    print("\\hline", file=f_)
    print(f"Code promised & && \\\\", file=f_)
    print("(according to reviewers) $\\rightarrow$ & No info & Yes  &Total\\\\", file=f_)
    print("&&&\\\\", file=f_)
    print("Code available & && \\\\(after acceptance) $\downarrow$ & && \\\\", file=f_)
    print("\\hline", file=f_)
    c_=df_here.iloc[1].to_numpy()
    c2_=df_here2.iloc[1].to_numpy()
    print(f" No repo & {float(c_[1]):.0f} ({float(c2_[1]):.0f}\%)& {float(c_[2]):.0f} ({float(c2_[2]):.0f}\%)& {float(c_[3]):.0f} ({float(c2_[3]):.0f}\%)\\\\", file=f_)
    c_=df_here.iloc[0].to_numpy()
    c2_=df_here2.iloc[0].to_numpy()
    print(f" Invalid repo & {float(c_[1]):.0f} ({float(c2_[1]):.0f}\%)& {float(c_[2]):.0f} ({float(c2_[2]):.0f}\%)& {float(c_[3]):.0f} ({float(c2_[3]):.0f}\%)\\\\", file=f_)
    c_=df_here.iloc[2].to_numpy()
    c2_=df_here2.iloc[2].to_numpy()
    print(f" Valid repo & {float(c_[1]):.0f} ({float(c2_[1]):.0f}\%)& {float(c_[2]):.0f} ({float(c2_[2]):.0f}\%)& {float(c_[3]):.0f} ({float(c2_[3]):.0f}\%)\\\\", file=f_)

    c_=df_here.iloc[3].to_numpy()
    c2_=df_here2.iloc[3].to_numpy()
    print(f" Total & {float(c_[1]):.0f} ({float(c2_[1]):.0f}\%)& {float(c_[2]):.0f} ({float(c2_[2]):.0f}\%)& {float(c_[3]):.0f} ({float(c2_[3]):.0f}\%)\\\\", file=f_)
    print("\\hline", file=f_)
    print("\\end{tabular}", file=f_)

print(f"Wrote tex file {latex_directory / 'table_code_availability.tex'}")

### Word counts

df_stats_word_counts = pd.read_csv(output_directory / "stats" / "stats_word_counts.csv", sep= "\t")
print(tabulate(df_stats_word_counts, headers='keys', tablefmt='psql'))

latex_directory = output_directory / "latex"
if not latex_directory.is_dir():
    latex_directory.mkdir()

with open(latex_directory / 'table_S_nbwords.tex', 'w') as f_:
    print("\\begin{tabular}{lllllll}", file=f_)
    print("\\hline", file=f_)
    print(f"Category & Mean & Min & Max & Median & $\sigma$ & IQR \\\\ ", file=f_)
    print("\\hline", file=f_)
    c_=df_stats_word_counts.iloc[0].to_numpy()
    print(f"Contribution & {c_[1]:.2f} & {c_[3]:.0f} & {c_[2]:.0f} & {c_[4]:.0f} & {c_[5]:.2f} & {c_[6]:.0f} \\\\", file=f_)
    c_=df_stats_word_counts.iloc[1].to_numpy()
    print(f"Strengths & {c_[1]:.2f} & {c_[3]:.0f} & {c_[2]:.0f} & {c_[4]:.0f} & {c_[5]:.2f} & {c_[6]:.0f} \\\\", file=f_)
    c_=df_stats_word_counts.iloc[2].to_numpy()
    print(f"Weaknesses & {c_[1]:.2f} & {c_[3]:.0f} & {c_[2]:.0f} & {c_[4]:.0f} & {c_[5]:.2f} & {c_[6]:.0f} \\\\", file=f_)
    c_=df_stats_word_counts.iloc[3].to_numpy()
    print(f"Reproducibility & {c_[1]:.2f} & {c_[3]:.0f} & {c_[2]:.0f} & {c_[4]:.0f} & {c_[5]:.2f} & {c_[6]:.0f} \\\\", file=f_)
    c_=df_stats_word_counts.iloc[4].to_numpy()
    print(f"Detailed comments & {c_[1]:.2f} & {c_[3]:.0f} & {c_[2]:.0f} & {c_[4]:.0f} & {c_[5]:.2f} & {c_[6]:.0f} \\\\", file=f_)
    c_=df_stats_word_counts.iloc[5].to_numpy()
    print(f"Justification & {c_[1]:.2f} & {c_[3]:.0f} & {c_[2]:.0f} & {c_[4]:.0f} & {c_[5]:.2f} & {c_[6]:.0f} \\\\", file=f_)
    print("\\hline", file=f_)
    c_=df_stats_word_counts.iloc[6].to_numpy()
    print(f"Total & {c_[1]:.2f} & {c_[3]:.0f} & {c_[2]:.0f} & {c_[4]:.0f} & {c_[5]:.2f} & {c_[6]:.0f} \\\\", file=f_)
    print("\\hline", file=f_)
    print("\\end{tabular}", file=f_)

print(f"Wrote tex file {latex_directory / 'table_S_nbwords.tex'}")

### Suppl: statement

with open(output_directory / "stats_rating" / "data.json", 'r') as file:
    data_json = json.load(file)

nb_reviews = data_json['nb_reviews']
print(f"Number of Papers: {data_json['nb_papers']}")
print(f"Number of Reviews: {nb_reviews}")

df_inter = pd.read_csv(output_directory / "stats_rating" / "2-statements.csv", sep= ";")
print(tabulate(df_inter, headers='keys', tablefmt='psql'))

latex_directory = output_directory / "latex"
if not latex_directory.is_dir():
    latex_directory.mkdir()

with open(latex_directory / 'table_S_statement.tex', 'w') as f_:
    print("\\begin{tabular}{llll}", file=f_)
    print("\\hline", file=f_)
    print(f" & \% & 95\% CI  & N \\\\", file=f_)
    print("\\hline", file=f_)
    c_=df_inter.iloc[0].to_numpy()
    print(f"(+) & {c_[2]:.0f}\% & [{c_[3]:.0f}\%, {c_[4]:.0f}\%] & {c_[1]:.0f}/{nb_reviews}\\\\", file=f_)
    c_=df_inter.iloc[1].to_numpy()
    print(f"($-$) & {c_[2]:.0f}\% & [{c_[3]:.0f}\%, {c_[4]:.0f}\%] & {c_[1]:.0f}/{nb_reviews}\\\\", file=f_)
    c_=df_inter.iloc[2].to_numpy()
    print(f"(none) & {c_[2]:.0f}\% & [{c_[3]:.0f}\%, {c_[4]:.0f}\%] & {c_[1]:.0f}/{nb_reviews}\\\\", file=f_)
    c_=df_inter.iloc[3].to_numpy()
    print(f"(unusable) & {c_[2]:.0f}\% & [{c_[3]:.0f}\%, {c_[4]:.0f}\%] & {c_[1]:.0f}/{nb_reviews}\\\\", file=f_)
    print("\\hline", file=f_)
    print("\\end{tabular}", file=f_)

print(f"Wrote tex file {latex_directory / 'table_S_statement.tex'}")

### Suppl: comments

with open(output_directory / "stats_rating" / "data.json", 'r') as file:
    data_json = json.load(file)

nb_reviews = data_json['nb_reviews']
print(f"Number of Papers: {data_json['nb_papers']}")
print(f"Number of Reviews: {nb_reviews}")

df_inter = pd.read_csv(output_directory / "stats_rating" / "3-comments.csv", sep= ";")
print(tabulate(df_inter, headers='keys', tablefmt='psql'))

latex_directory = output_directory / "latex"
if not latex_directory.is_dir():
    latex_directory.mkdir()

with open(latex_directory / 'table_S_comments.tex', 'w') as f_:
    print("\\begin{tabular}{llll}", file=f_)
    print("\\hline", file=f_)
    print(f"& \% & 95\% CI  & N \\\\", file=f_)
    print("\\hline", file=f_)
    c_=df_inter.iloc[1].to_numpy()
    print(f"(+) & {c_[2]:.0f}\% & [{c_[3]:.0f}\%, {c_[4]:.0f}\%] & {c_[1]:.0f}/{nb_reviews}\\\\", file=f_)
    c_=df_inter.iloc[2].to_numpy()
    print(f"($-$) & {c_[2]:.0f}\% & [{c_[3]:.0f}\%, {c_[4]:.0f}\%] & {c_[1]:.0f}/{nb_reviews}\\\\", file=f_)
    c_=df_inter.iloc[0].to_numpy()
    print(f"($-/+$) & {c_[2]:.0f}\% & [{c_[3]:.0f}\%, {c_[4]:.0f}\%] & {c_[1]:.0f}/{nb_reviews}\\\\", file=f_)
    c_=df_inter.iloc[3].to_numpy()
    print(f"(none) & {c_[2]:.0f}\% & [{c_[3]:.0f}\%, {c_[4]:.0f}\%] & {c_[1]:.0f}/{nb_reviews}\\\\", file=f_)
    c_=df_inter.iloc[4].to_numpy()
    print(f"(unusable) & {c_[2]:.0f}\% & [{c_[3]:.0f}\%, {c_[4]:.0f}\%] & {c_[1]:.0f}/{nb_reviews}\\\\", file=f_)
    print("\\hline", file=f_)
    print("\\end{tabular}", file=f_)

print(f"Wrote tex file {latex_directory / 'table_S_comments.tex'}")

### Suppl: meta-category

with open(output_directory / "stats_rating" / "data.json", 'r') as file:
    data_json = json.load(file)

nb_reviews = data_json['nb_reviews']
print(f"Number of Papers: {data_json['nb_papers']}")
print(f"Number of Reviews: {nb_reviews}")

df_inter = pd.read_csv(output_directory / "stats_rating" / "5-meta.csv", sep= ";")
print(tabulate(df_inter, headers='keys', tablefmt='psql'))

latex_directory = output_directory / "latex"
if not latex_directory.is_dir():
    latex_directory.mkdir()

with open(latex_directory / 'table_S_meta.tex', 'w') as f_:
    print("\\begin{tabular}{llll}", file=f_)
    print("\\hline", file=f_)
    print(f"& \% & 95\% CI  & N \\\\", file=f_)
    print("\\hline", file=f_)
    c_=df_inter.iloc[0].to_numpy()
    print(f"(+) & {c_[2]:.0f}\% & [{c_[3]:.0f}\%, {c_[4]:.0f}\%] & {c_[1]:.0f}/{nb_reviews}\\\\", file=f_)
    c_=df_inter.iloc[1].to_numpy()
    print(f"($-$) & {c_[2]:.0f}\% & [{c_[3]:.0f}\%, {c_[4]:.0f}\%] & {c_[1]:.0f}/{nb_reviews}\\\\", file=f_)
    c_=df_inter.iloc[2].to_numpy()
    print(f"(unusable) & {c_[2]:.0f}\% & [{c_[3]:.0f}\%, {c_[4]:.0f}\%] & {c_[1]:.0f}/{nb_reviews}\\\\", file=f_)
    print("\\hline", file=f_)
    print("\\end{tabular}", file=f_)

print(f"Wrote tex file {latex_directory / 'table_S_meta.tex'}")

### Suppl: statements VS comments

with open(output_directory / "stats_rating" / "data.json", 'r') as file:
    data_json = json.load(file)

nb_reviews = data_json['nb_reviews']
print(f"Number of Papers: {data_json['nb_papers']}")
print(f"Number of Reviews: {nb_reviews}")

df_inter = pd.read_csv(output_directory / "stats_rating" / "4-statements_comments.csv", sep= ";")
print(tabulate(df_inter, headers='keys', tablefmt='psql'))

latex_directory = output_directory / "latex"
if not latex_directory.is_dir():
    latex_directory.mkdir()

with open(latex_directory / 'table_S_statement_comments.tex', 'w') as f_:
    print("\\begin{tabular}{lllllll}", file=f_)
    print("\\hline", file=f_)
    print(f"Comments $\\rightarrow$ & (+) & ($-$) & ($-/+$) & (none) &(unusable) & Total \\\\", file=f_)
    print(f"Statement $\\downarrow$ & &&&&& \\\\", file=f_)
    print("\\hline", file=f_)
    c_=df_inter.iloc[0].to_numpy()
    print(f"(+) & {c_[2]:.0f} & {c_[3]:.0f} & {c_[1]:.0f} & {c_[4]:.0f} & {c_[5]:.0f}& {c_[6]:.0f}\\\\", file=f_)
    c_=df_inter.iloc[1].to_numpy()
    print(f"($-$) & {c_[2]:.0f} & {c_[3]:.0f} & {c_[1]:.0f} & {c_[4]:.0f} & {c_[5]:.0f}& {c_[6]:.0f}\\\\", file=f_)
    c_=df_inter.iloc[2].to_numpy()
    print(f"(none) & {c_[2]:.0f} & {c_[3]:.0f} & {c_[1]:.0f} & {c_[4]:.0f} & {c_[5]:.0f}& {c_[6]:.0f}\\\\", file=f_)
    c_=df_inter.iloc[3].to_numpy()
    print(f"(unusable) & {c_[2]:.0f} & {c_[3]:.0f} & {c_[1]:.0f} & {c_[4]:.0f} & {c_[5]:.0f}& {c_[6]:.0f}\\\\", file=f_)
    c_=df_inter.iloc[4].to_numpy()
    print(f"Total & {c_[2]:.0f} & {c_[3]:.0f} & {c_[1]:.0f} & {c_[4]:.0f} & {c_[5]:.0f}& {c_[6]:.0f}\\\\", file=f_)
    print("\\hline", file=f_)
    print("\\end{tabular}", file=f_)

print(f"Wrote tex file {latex_directory / 'table_S_statement_comments.tex'}")
